requirements.txt
google-generativeai>=0.3.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.4.2
chromadb>=0.4.22
hnswlib>=0.7.0 